+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - More reliably handle and report diagnostics for scripts with invalid configurations. Specifically fixed https://github.com/google/wireit/issues/803.
 
+- Gracefully handle errors from the GitHub download cache API.
+
 ## [0.13.0] - 2023-09-01
 
 ### Changed